PO-2659: Refactor exception handling for Major Creditor API to improve error response structure
