+++ src/integrationTest/java/uk/gov/hmcts/opal/controllers/LegacyMinorCreditorPatchIntegrationTest.java
@@ -194,7 +194,7 @@ class LegacyMinorCreditorPatchIntegrationTest extends MinorCreditorControllerInt
             HttpStatus.SERVICE_UNAVAILABLE, "Service Unavailable", HttpHeaders.EMPTY, null, null));
 
         performLegacyPatch(503L, "\"1\"")
-            .andExpect(status().isServiceUnavailable())
+            .andExpect(status().isInternalServerError())
             .andExpect(content().contentType(MediaType.APPLICATION_PROBLEM_JSON_VALUE));
     }
